flask_app/requirements.txt

Flask>=2.0,<3.0
tensorflow
pillow
numpy
